% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Perform the process for each of the values in the ns vector and return a data frame with the results}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{A vector that will be sampled from}

\item{reps}{how many times many_sample_means will be performed}

\item{ns}{A vector representing the different values of n}
}
\value{
A data frame with the results
}
\description{
Perform the process for each of the values in the ns vector and return a data frame with the results
}
